+++ README.md
@@ -1,2 +1,12 @@
-# cryptocurrency-datamining
-cryptocurrency data analysis using Binance API
+# Notice: this is an alpha version for educational purposes
+
+### Features
+
+- Cryptocurrency data analysis using Binance API
+- Use of TensorFlow 2.10 , Keras and sklearn
+- Use of Python 3.9.0
+
+### Links
+This work is based on MEET NAGADIA notebook in kaggle. Please visit:
+
+`<link>` : <https://www.kaggle.com/code/meetnagadia/bitcoin-price-prediction-using-lstm>
