+++ img_output/README.md
@@ -1,0 +1,1 @@
+keep this folder, the project uses it to store output
